major improvement:
- multiple microscopic environments can be now simulated in the
multi-scale scenario (see StatenIslandRunner.generateScenario() and
StatenIslandRunner.runScenario() for details) and their origin
coordinate and orientation can be configured. 
For the time being the management of coordinate systems is still
missing. 

- TacticalDestination instances can be drawn in the .csv environment
file with linear shapes and arbitrary thickness. 

- Improved the visualization with the rotation of the network, according
to the origin coordinate and rotation configured. The visualization of
the microscopic environment cannot be rotated yet, due to missing
methods for the rotation of rectangles to be drawn (it is assumed to
draw it with no rotation).