Added JSONScenarioGenerator that creates a hybrid simulation by loading geometries and network info from JSON files.
Modified AlternativeBruteForceMultiScaleManger, defining an error
function with output generally ranging in [0:1] (it can be more than 1
but it is very rare and it indicates a very bad calibration). Still the
iterative process is not stopped after the network has reached a good
calibration (the stop criterion is defined but it does not halt the
system).

pom.xml is modified to import a library that parse json files